requirements.txt
streamlit
scikit-learn
joblib
nltk
PyPDF2
python-docx
numpy